feat(backend): refactor endpoint to manage rounded flights
