+++ backend/src/main/java/encora/breakableII/backend/models/FlightOffer.java
@@ -1,25 +1,23 @@
 package encora.breakableII.backend.models;
 
+import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public class FlightOffer {
-    private LocalDateTime departureDate;
-    private LocalDateTime arrivalDateTime;
-    private String departureAirport;
-    private String departureAirportName;
-    private String arrivalAirport;
-    private String arrivalAirportName;
-    private String airlineName;
-    private String airlineCode;
-    private String carrierName;
-    private String carrierCode;
-    private String duration;
+    private List<IndividualFlight> individualFlights;
+    private String id;
     private String pricePerTraveler;
     private String priceTotal;
-    private String currency;
-    private List<Stops> stops;
-    private String id;
+    private Duration totalDuration;
+
+    public Duration getTotalDuration() {
+        return totalDuration;
+    }
+
+    public void setTotalDuration(Duration totalDuration) {
+        this.totalDuration = totalDuration;
+    }
 
     public String getId() {
         return id;
@@ -29,28 +27,12 @@ public class FlightOffer {
         this.id = id;
     }
 
-    public String getCurrency() {
-        return currency;
+    public List<IndividualFlight> getIndividualFlights() {
+        return individualFlights;
     }
 
-    public void setCurrency(String currency) {
-        this.currency = currency;
-    }
-
-    public String getDepartureAirportName() {
-        return departureAirportName;
-    }
-
-    public void setDepartureAirportName(String departureAirportName) {
-        this.departureAirportName = departureAirportName;
-    }
-
-    public String getArrivalAirportName() {
-        return arrivalAirportName;
-    }
-
-    public void setArrivalAirportName(String arrivalAirportName) {
-        this.arrivalAirportName = arrivalAirportName;
+    public void setIndividualFlights(List<IndividualFlight> individualFlights) {
+        this.individualFlights = individualFlights;
     }
 
     public String getPricePerTraveler() {
@@ -67,85 +49,5 @@ public class FlightOffer {
 
     public void setPriceTotal(String priceTotal) {
         this.priceTotal = priceTotal;
-    }
-
-    public String getDepartureAirport() {
-        return departureAirport;
-    }
-
-    public void setDepartureAirport(String departureAirport) {
-        this.departureAirport = departureAirport;
-    }
-
-    public String getArrivalAirport() {
-        return arrivalAirport;
-    }
-
-    public void setArrivalAirport(String arrivalAirport) {
-        this.arrivalAirport = arrivalAirport;
-    }
-
-    public LocalDateTime getDepartureDate() {
-        return departureDate;
-    }
-
-    public void setDepartureDate(LocalDateTime departureDate) {
-        this.departureDate = departureDate;
-    }
-
-    public LocalDateTime getArrivalDateTime() {
-        return arrivalDateTime;
-    }
-
-    public void setArrivalDateTime(LocalDateTime arrivalDateTime) {
-        this.arrivalDateTime = arrivalDateTime;
-    }
-
-    public String getAirlineName() {
-        return airlineName;
-    }
-
-    public void setAirlineName(String airlineName) {
-        this.airlineName = airlineName;
-    }
-
-    public String getAirlineCode() {
-        return airlineCode;
-    }
-
-    public void setAirlineCode(String airlineCode) {
-        this.airlineCode = airlineCode;
-    }
-
-    public String getCarrierName() {
-        return carrierName;
-    }
-
-    public void setCarrierName(String carrierName) {
-        this.carrierName = carrierName;
-    }
-
-    public String getCarrierCode() {
-        return carrierCode;
-    }
-
-    public void setCarrierCode(String carrierCode) {
-        this.carrierCode = carrierCode;
-    }
-
-    public String getDuration() {
-        return duration;
-    }
-
-    public void setDuration(String duration) {
-        this.duration = duration;
-    }
-
-    public List<Stops> getStops() {
-        return stops;
-    }
-
-    public void setStops(List<Stops> stops) {
-        this.stops = stops;
     }
 }
